refactor: BaseException.java 추가 및 명칭변경
