+++ src/main/java/com/hhplus/architecture/domain/lectureRegistration/LecturesRegisteredByStudent.java
@@ -4,7 +4,7 @@ import com.hhplus.architecture.domain.lecture.LectureInfo;
 import com.hhplus.architecture.domain.student.StudentInfo;
 import java.util.List;
 
-public record LectureRegistrationInfo(
+public record LecturesRegisteredByStudent(
     StudentInfo studentInfo,
     List<LectureInfo> lectureInfos
 ) {
