+++ src/main/java/com/hhplus/architecture/support/BaseControllerAdvice.java
@@ -1,24 +1,22 @@
-package com.hhplus.architecture.api.lecture;
+package com.hhplus.architecture.support;
 
-import com.hhplus.architecture.domain.lecture.LectureException;
-import com.hhplus.architecture.domain.lecture.LectureException.ErrorCode;
-import com.hhplus.architecture.support.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 @RestControllerAdvice
-class LectureControllerAdvice extends ResponseEntityExceptionHandler {
+class BaseControllerAdvice extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(value = Exception.class)
     public ResponseEntity<ErrorResponse> handleException(Exception e) {
         return ResponseEntity.status(500).body(new ErrorResponse("500", "에러가 발생했습니다."));
     }
 
-    @ExceptionHandler(value = LectureException.class)
-    public ResponseEntity<ErrorResponse> handleException(LectureException e) {
-        ErrorCode errorCode = e.getErrorCode();
+    @ExceptionHandler(value = BaseException.class)
+    public ResponseEntity<ErrorResponse> handleException(BaseException e) {
+        BaseErrorCode errorCode = e.getErrorCode();
         return ResponseEntity.status(errorCode.getStatus()).body(ErrorResponse.error(errorCode));
     }
+
 }
