+++ src/main/java/com/hhplus/architecture/support/BaseException.java
@@ -1,0 +1,15 @@
+package com.hhplus.architecture.support;
+
+public class BaseException extends RuntimeException {
+
+    private final BaseErrorCode errorCode;
+
+    public BaseException(BaseErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    public BaseErrorCode getErrorCode() {
+        return errorCode;
+    }
+}
